shogun! (please read comment)
Finally could compile the hell out of it but it needed a stupid amount
of libraries, including two that are outdated in the CentOS repos and
need to be downloaded and compiled from source: SWIG and EIGEN3. Also,
it searches for Lapack in Atlas even though it also has its own package,
so cmake blows a gasket because it sees duplicates everywhere, but just
ignore that, it works regardless. Have fun. :)